builder->jumper bugfix, digger physics fit image
There was a glitch involving jumper assignments on builders.
If you assigned right after laying the brick, the lix was inside
the terrain, and thus the jumper was immediately stunned.

The digger changed physics to a large extent. It will leave
a small ledge when digging out of a horizontal ceiling.