built time must match exactly to allow highscore
When Clam made levels, he was ahead of our local time due to time zones.
I played the levels and got highscores. When I altered the levels, they were
older instead of newer, so my highscores were still vaild. I don't want this,
so now the "built" of the level must match the required "built" of the score.