level hints, read doc/files.txt, bug: hint overwrites dialog
The Lix editor doesn't alter level hints, use a text editor on the level file.
There, hints are defined as a single-line string, with `<br>' for line breaks.
Hints cannot be deactivated, the intention is to use them in tutorial levels.
Prefer a lower y length in such levels, or ensure a neutral background.

There's an unresolved bug, levels with large hints erase the top part of
the ESC/end-of-level menu.