bugfix: ascender in midair, version to 2014-03-21
In the tumbler code, there was a default value of 0 for "no floor here",
but that could have been a floor at a certain pixel as well. Default value
is now -999, which is different from every checked height. Bug before:
Climbable jumper hit top of a block, then ascended way above the block.