bugfix: digger was cancellable with blocker
The digger had one frame in mid-air, even without having dug through
the terrain already. The digger now moves down a frame ealier as a fix.
The spritesheet hasn't been updated yet, so set_y is called on the sprite.
This can be removed at once when the spritesheet gets updated in a future
fix/feature.