directional force transmitted by replays and network
This shall remedy the following bug: In a multiplayer team game, if both
players e.g. assign a walker to the same lix to turn her around, two walkers
will be assigned and the lix will end up walking in the same direction as
before. If both players use directional force, this could be amended if the
directional force option was transmitted as well, and the assignment is only
done if the direction of the lix fits. So the later of the two assignments
will be invalid as the directions don't match.