bugfix: bracketed checkmark didn't vanish
Description: I was experimenting with a change to a level. I decided against
it, so I used geoo's replay and the "Extract level" function to retrieve the
previous version of the level. When I went back into the level select menu,
I got the bracketed tick and "Level has been edited" in place of the usual
statistics -- and they remained even after I completed the level again.

Solution: solving a level makes the result always better if
solve.built != old.built, not just solve.built > old.built. Here, built is
the date when the level was built.