bugfix: editor exit -> save -> lockup
This fixes the following bug from the bugtracker:

Go into a directory in which no level is selected; press Edit to start a new
blank level; make some changes; press the "exit editor" button. Lix will ask
you if you want to create a new level. When you enter the filename and press
Enter, Lix crashes. Or instead of crashing, it gets stuck and the editor
can't be used anymore.