escape space, dash, singlequote; user loading fallback
No matter how many ASCII chars we don't escape, we'd have to
implement a fallback in other/user.cpp anyway, like the one
in this commit. Therefore, we go for a super-simple rule on
what chars will be escaped: All except A-Z a-z 0-9 will be
escaped. Simplicity is a design goal.